+++ src/mlpack/methods/ann/layer/flexible_relu.hpp
@@ -115,11 +115,6 @@ class FlexibleReLU
   //! Modify the parameters.
   OutputDataType& Parameters() { return alpha; }
 
-  //! Get the input parameter.
-  InputDataType const& InputParameter() const { return inputParameter; }
-  //! Modify the input parameter.
-  InputDataType& InputParameter() { return inputParameter; }
-
   //! Get the output parameter.
   OutputDataType const& OutputParameter() const { return outputParameter; }
   //! Modify the output parameter.
@@ -149,9 +144,6 @@ class FlexibleReLU
  private:
   //! Locally-stored delta object.
   OutputDataType delta;
-
-  //! Locally-stored input parameter object.
-  InputDataType inputParameter;
 
   //! Locally-stored output parameter object.
   OutputDataType outputParameter;