+++ src/mlpack/methods/cf/neighbor_search_policies/pearson_search.hpp
@@ -68,7 +68,7 @@ class PearsonSearch
     // For each vector x, first subtract mean(x) from each element in x.
     // Then normalize the vector to unit length.
     arma::mat normalizedQuery;
-    normalizedQuery = arma::normalise(query - arma::mean(query));
+    normalizedQuery = arma::normalise(query.each_row() - arma::mean(query));
 
     neighborSearch.Search(normalizedQuery, k, neighbors, similarities);